services/image_captioning/requirements.txt
transformers
torch
pillow
accelerate
timm 
uvicorn 
python-multipart
fastapi